fix: split duplicate to components
